Break down big file
